requirements.txt
numpydoc==1.1.0
pytest==6.1.2
pylint==2.6.0
sphinx==3.3.1
sphinx-rtd-theme==0.5.0
XInput-Python==0.4.0
pylsl==1.14.0
psycopg2==2.8.6
pyserial==3.1.1